schema: extract Manifest into external package
