+++ go.mod
@@ -8,7 +8,11 @@ require (
 	github.com/go-sql-driver/mysql v1.3.0
 	github.com/gorilla/mux v1.7.0
 	github.com/jessevdk/go-flags v1.4.0
-	github.com/mattn/go-sqlite3 v1.10.0 // indirect
-	github.com/uoregon-libraries/gopkg v0.15.0
+	github.com/uoregon-libraries/gopkg v0.18.0
 	golang.org/x/crypto v0.0.0-20210711020723-a769d52b0f97
+)
+
+require (
+	github.com/mattn/go-sqlite3 v1.10.0 // indirect
+	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
 )
