--- src/schema/testdata/.gitignore
@@ -1,1 +1,0 @@
-/.manifest
