--- src/schema/testdata/a.txt
@@ -1,1 +1,0 @@
-this is the file known as "a"
